requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
djangorestframework==3.14.0
everapi==0.1.1
idna==3.4
inflection==0.5.1
packaging==23.1
psycopg2-binary==2.9.7
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
